# docs/swagger.yaml
basePath: /api/v1
definitions:
  POS-kasir_internal_common.ErrorResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
    type: object
  POS-kasir_internal_common.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  POS-kasir_internal_common_pagination.Pagination:
    properties:
      current_page:
        type: integer
      per_page:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
...
  POS-kasir_internal_dto.ApplyPromotionRequest:
    properties:
      promotion_id:
        type: string
    required:
    - promotion_id
    type: object
...
  POS-kasir_internal_dto.CancelOrderRequest:
    properties:
      cancellation_notes:
        maxLength: 255
        type: string
      cancellation_reason_id:
        type: integer
    required:
    - cancellation_reason_id
    type: object
...
  POS-kasir_internal_dto.ConfirmManualPaymentRequest:
    properties:
      cash_received:
        minimum: 0
        type: integer
      payment_method_id:
        type: integer
    required:
    - payment_method_id
    type: object
...
  POS-kasir_internal_dto.CreateOrderItemRequest:
    properties:
      options:
        items:
          $ref: '#/definitions/POS-kasir_internal_dto.CreateOrderItemOptionRequest'
        type: array
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  POS-kasir_internal_dto.CreateOrderRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/POS-kasir_internal_dto.CreateOrderItemRequest'
        minItems: 1
        type: array
      type:
        allOf:
        - $ref: '#/definitions/POS-kasir_internal_repository.OrderType'
        enum:
        - dine_in
        - takeaway
    required:
    - items
    - type
    type: object
...
  POS-kasir_internal_dto.MidtransPaymentResponse:
    properties:
      actions:
        items:
          $ref: '#/definitions/POS-kasir_internal_dto.PaymentAction'
        type: array
      expiry_time:
        type: string
      gross_amount:
        type: string
      order_id:
        type: string
      qr_string:
        type: string
      transaction_id:
        type: string
    type: object
  POS-kasir_internal_dto.OrderDetailResponse:
    properties:
      applied_promotion_id:
        type: string
      cash_received:
        type: integer
      change_due:
        type: integer
      created_at:
        type: string
      discount_amount:
        type: integer
      gross_total:
        type: integer
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/POS-kasir_internal_dto.OrderItemResponse'
        type: array
      net_total:
        type: integer
      payment_gateway_reference:
        type: string
      payment_method_id:
        type: integer
      status:
        $ref: '#/definitions/POS-kasir_internal_repository.OrderStatus'
      type:
        $ref: '#/definitions/POS-kasir_internal_repository.OrderType'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  POS-kasir_internal_dto.OrderItemResponse:
    properties:
      id:
        type: string
      options:
        items:
          $ref: '#/definitions/POS-kasir_internal_dto.OrderItemOptionResponse'
        type: array
      price_at_sale:
        type: integer
      product_id:
        type: string
      product_name:
        type: string
      quantity:
        type: integer
      subtotal:
        type: integer
    type: object
  POS-kasir_internal_dto.OrderListResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_paid:
        type: boolean
      items:
        items:
          $ref: '#/definitions/POS-kasir_internal_dto.OrderItemResponse'
        type: array
      net_total:
        type: integer
      queue_number:
        type: string
      status:
        $ref: '#/definitions/POS-kasir_internal_repository.OrderStatus'
      type:
        $ref: '#/definitions/POS-kasir_internal_repository.OrderType'
      user_id:
        type: string
    type: object
  POS-kasir_internal_dto.PagedOrderResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/POS-kasir_internal_dto.OrderListResponse'
        type: array
      pagination:
        $ref: '#/definitions/POS-kasir_internal_common_pagination.Pagination'
    type: object
...
  POS-kasir_internal_dto.UpdateOrderItemRequest:
    properties:
      options:
        items:
          $ref: '#/definitions/POS-kasir_internal_dto.CreateOrderItemOptionRequest'
        type: array
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  POS-kasir_internal_dto.UpdateOrderStatusRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/POS-kasir_internal_repository.OrderStatus'
        enum:
        - in_progress
        - served
        - paid
    required:
    - status
    type: object
...
  POS-kasir_internal_repository.OrderType:
    enum:
    - dine_in
    - takeaway
    type: string
    x-enum-varnames:
    - OrderTypeDineIn
    - OrderTypeTakeaway
...
host: localhost:8080
info:
  contact: {}
  description: POS Kasir API
  title: POS Kasir API
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of orders per page
        in: query
        name: limit
        type: integer
      - collectionFormat: multi
        description: Order statuses
        in: query
        items:
          enum:
          - open
          - in_progress
          - served
          - paid
          - cancelled
          type: string
        name: statuses
        type: array
      - description: Filter by User ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/POS-kasir_internal_dto.PagedOrderResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: List orders
      tags:
      - Orders
      x-roles:
      - admin
      - manager
      - cashier
    post:
      consumes:
      - application/json
      parameters:
      - description: Create order details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/POS-kasir_internal_dto.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/POS-kasir_internal_dto.OrderDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Create an order
      tags:
      - Orders
      x-roles:
      - admin
      - manager
      - cashier
  /orders/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/POS-kasir_internal_dto.OrderDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Get an order by ID
      tags:
      - Orders
      x-roles:
      - admin
      - manager
      - cashier
  /orders/{id}/apply-promotion:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Promotion details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/POS-kasir_internal_dto.ApplyPromotionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/POS-kasir_internal_dto.OrderDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Apply promotion to an order
      tags:
      - Orders
      x-roles:
      - admin
      - manager
      - cashier
  /orders/{id}/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Cancel order details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/POS-kasir_internal_dto.CancelOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Cancel an order
      tags:
      - Orders
      x-roles:
      - admin
      - manager
      - cashier
  /orders/{id}/items:
    put:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Update order items
        in: body
        name: request
        required: true
        schema:
          items:
            $ref: '#/definitions/POS-kasir_internal_dto.UpdateOrderItemRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/POS-kasir_internal_dto.OrderDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Update items in an order
      tags:
      - Orders
      x-roles:
      - admin
      - manager
      - cashier
  /orders/{id}/pay/manual:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Manual payment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/POS-kasir_internal_dto.ConfirmManualPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/POS-kasir_internal_dto.OrderDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Confirm manual payment for an order
      tags:
      - Orders
      x-roles:
      - admin
      - manager
      - cashier
  /orders/{id}/pay/midtrans:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/POS-kasir_internal_dto.MidtransPaymentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Initiate Midtrans payment for an order
      tags:
      - Orders
      x-roles:
      - admin
      - manager
      - cashier
...
  /orders/{id}/update-status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Order status details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/POS-kasir_internal_dto.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/POS-kasir_internal_common.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/POS-kasir_internal_dto.OrderDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/POS-kasir_internal_common.ErrorResponse'
      summary: Update order operational status
      tags:
      - Orders
      x-roles:
      - admin
      - manager
      - cashier
...
swagger: "2.0"